ref: split and rename smtp test so it actually runs
